reloading the map each time something changed in Core Data and fixing status bar
Before : The list of Location objects is only fetched once in
viewDidLoad(), any changes that happen afterwards are overlooked.

Solution : Using the NotificationCenter for the Map screen. Core Data
sends out a bunch of notifications when changes are made to the data
store. Subscribe to these notifications and update the map view when
you receive them.

- Fix the issue with the status bar